Throw an error if Tx serialisation fails
